refactor: poem 충돌 이슈 해결
